Refactor title screen and allow changing the seed
