+++ src/main.ts
@@ -2,24 +2,10 @@ import './cssReset.css';
 import './style.css';
 import { Builder } from './components/Builder/Builder';
 import { DynamicScreenMaker } from './components/Screens/DynamicScreenMaker';
+import { GenerateTitleScreen } from './components/Utilities/GenerateTitleScreen';
 
 const SHOW_MENU: boolean = true;
-const seed: number = 99;
-
-const showTitleScreen = () => {
-  const titleScreen = document.createElement('title-screen');
-  const titleContainer = document.createElement('div');
-  titleContainer.id = 'title-container';
-  titleContainer.appendChild(titleScreen);
-
-  document.body.insertBefore(titleContainer, document.body.firstChild);
-
-  titleScreen.dispatchEvent(new CustomEvent('pass-seed', { detail: { seed } }));
-  titleScreen.addEventListener('start-new-game', () => {
-    titleContainer.remove();
-    DynamicScreenMaker.runBuilt_InitialGameSetup(new Builder(seed), seed);
-  });
-};
+const seed: number = 1337;
 
 const runGameDirectly = () => {
   document.addEventListener('DOMContentLoaded', () => {
@@ -32,7 +18,7 @@ const runGameDirectly = () => {
 };
 
 if (SHOW_MENU) {
-  showTitleScreen();
+  GenerateTitleScreen.generate(seed);
 } else {
   runGameDirectly();
 }
